requirements.txt
streamlit
chromadb
ollama
pypdf
python-dotenv
google-generativeai
langchain
